+++ raspberry-pi/app/motors/__init__.py
@@ -1,0 +1,5 @@
+"""Hardware-separated motor interfaces."""
+
+from .stepper import LgpioStepper, StepperError
+
+__all__ = ["LgpioStepper", "StepperError"]